feat: add desktop-window-utils for window utilities
