+++ packages/desktop-window/src/index.ts
@@ -9,3 +9,4 @@ export * from './lib/desktop-window-setup';
 export * from './lib/desktop-window-timer';
 export * from './lib/desktop-window-updater';
 export * from './lib/screen-capture-notification';
+export * from './lib/utils/desktop-window-utils';
